[carnelian] Add useful arguments to infinite scroll example.
This add --max-column=NUMBER and --disable-text arguments to
this example. They make it possible to use this example for input
processing tests on devices that can't render the default
contents at 60 FPS.

Also avoids making a full redraw in CopyRedraw mode when scroll
position has not changed.

Touch event timestamp logic was also updated to support driver
timestamps and print a warning if the touch driver is not
providing timestamps.

Test: run fuchsia-pkg://fuchsia.com/infinite_scroll_rs#meta/infinite_scroll_rs.cmx --disable-text --max-columns=1 --scale=0.25
Change-Id: I242fd9dcc38c6c8669ed35b39f7319df6fe126b6