[audio_core] Use std::shared_ptr for Mixer.
As we look towards adding multiple mix stages, we need to begin to
decouple the Mixer from the AudioLink. There will be only 1 AudioLink
between a Renderer and an Output, but there may be multiple mix
stages/Mixer instances.

The AudioLink still needs access to the Mixer in order to set the
stream gain for now, so we use a std::shared_ptr.

Bug: 36279
Bug: 39624
Test: Refactor only; no behavior change
      fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
Change-Id: I450aef47d4bf406d5800893c83833339405bdad5