[sl4f] Gracefully handle hung diagnostics
The dart test binary was hanging because of stream subscriptions hidden
behind standard library calls that would not be cancelled in the case
where the diagnostic timeout was hit and we killed the process.

It is still unclear why this occurred on infra but not locally.

Tested to successfully exit on both infra and local, including with a
diagnostic command that infinite loops.

Bug: 42124

Change-Id: Ia6b500a48eb1c10ffd3c6989f870f61fb82e9ce6