[audio_core] Remove refs to self in AudioCapturer
These references to self existed to prevent the FIDL thread from
deleting resources in use by the mix thread.

Since BeginShutdown() synchronizes with the mix thread, this is no
longer needed. It is also unsafe as fbl::RefPtr(raw) is a weak ptr
upgrade we assume always succeeds.

Change-Id: I1c44e07413549f0b754d7c4051017065ed2c2c60

---

This is part of a series of changes which remove AudioLink from AudioCore.
AudioLink requires an implicit distributed ownership model where
destination objects own sources, and sources own destinations, requiring us
to manually free objects.

The new feature pressuring this change is the addition of a
LoudnessTransform to links, which must be added to links by the RouteGraph.
The transform determines what loudness stages take effect for a given
output.

The replacement is the LinkMatrix, a single owner for all AudioObject
connections. RouteGraph will own the LinkMatrix and all AudioObjects.

Initially this refactor will not be strict about ownership, preferring
shared_ptrs.

Once some restructuring in audio core is done to ensure drop order
consistency across threads, more strictness can be imposed at compile
time.

Bug: 39624
Bug: 39426
Bug: 43514

TEST: no behaviour change; fx run-test audio_core_unittests

Change-Id: If9d6a5f6799ba33b259d68f451aa3aa9fc885626