[FDF][SystemSuspend] Have a configurable timeout
Currently system wide suspend times out after 30 sec. To test this
behaviour, the test waits for 30 sec and makes sure the system wide
suspend proceeds although a particular suspend hook fails to complete on
time. The 30 sec timeout is unacceptable in tests. This change makes the
timeout configurable so that a test can test the behaviour without
waiting for 30 sec.

Bug:39914

Test: fx run-tests coordinator_tests.

Change-Id: Ie585f8cfacd2c9a527e4b25f4716ac783e51e803