[build] Append non_rust_deps to regular deps
With the native Rust support in GN, everything can be passed as a deps
and GN does the right thing depending on the type and language of the
target, so we should no longer need non_rust_deps.

The uses of non_rust_deps throughout the codebase will be removed in
subsequent changes.

Bug: 43781
Change-Id: I09666823deaa0837f5cc37cce2c424839eebe439