[inspect] Move NodeHierarchy out of fuchsia-inspect for reuse on host
New crate:

- fuchsia-inspect-node-hierarchy: contains the NodeHierarchy and the
  trie property iterator.

LinkNodeDisposition and ArrayFormat ended up in fuchsia-inspect-node-hierarchy
as both crates need it. This doesn't really belong here and should be in some
other crate of the form "fuchsia-inspect-format".

As a consequence of this, inspect-formatter is now reusable in the host
as well.

TESTED=existing tests continue to pass

Change-Id: I0277e22b62f3e51e62335ccee11dd88c79e14eca