[audio_core] AudioLink endpoints as references
This exposes the endpoints of AudioLinks as references, instead of
RefPtrs. This will allow AudioLink to take endpoints as shared_ptrs
without changing the endpoint's API, and liberates the endpoints of
redundant null checks.

=

This is part of a series of changes which remove AudioLink from AudioCore.
AudioLink requires an implicit distributed ownership model where
destination objects own sources, and sources own destinations, requiring us
to manually free objects.

The new feature pressuring this change is the addition of a
LoudnessTransform to links, which must be added to links by the RouteGraph.
The transform determines what loudness stages take effect for a given
output.

The replacement is the LinkMatrix, a single owner for all AudioObject
connections. RouteGraph will own the LinkMatrix and all AudioObjects.

Initially this refactor will not be strict about ownership, preferring
shared_ptrs.

Once some restructuring in audio core is done to ensure drop order
consistency across threads, more strictness can be imposed at compile
time.

Bug: 39624
Bug: 39426
Bug: 43514

TEST: No behaviour change;
      tests updated fx run-test audio_core_unittests

Change-Id: Ib37a7cd6c7591a7f4cd9fc1a73a278ee0f4d3181