[kazoo] Support compiling out testonly syscalls
When a syscall is marked up with [testonly] and the GN variable
exclude_testonly_syscalls is set, the syscalls will be excluded from the
build. This is mapped to an --exclude flag in Kazoo.

This variable is not yet hooked up pending coming up with what build
configuration/setting we want to use to accomplish this. Currently some
testing syscalls in //zircon/syscalls/syscall.fidl are tagged with the
attribute to be excluded.

Bug: 38409
Bug: 33324
Change-Id: I7b3cfe6e809366419744216e459305a5777ff407