[component-manager] Use and Expose Runtime Directory Right Enforcement.
This change enables runtime enforcement for directory rights for use and
expose declarations in the component framework. This process is
relatively complex and so a short description is provided below for each
case.

1. Use Declarations
If a use declaration relates to using a directory, rights must always be
provided. In this case the offers will be traversed until an offer from
self is found (in which case rights must be specified) or an
intermediate offer with rights is found. It will verify that the use
declaration is less than or equal to the closest parent offers rights.
From here the chain will be followed all the way back to the originating
source which can be a Framework, Bultin or Self node. At each hop in the
chain if rights are specified it will verify they are less than or equal
to the prior rights node in the chain.

2. Expose Declarations
If exposed from self the directory enforcement will attempt to open with
whatever flags it is provided and rely on the underlying fio operation
to deny access to privs it cannot obtain (for example attempting to mark
a random exposed directory as executable). If the expose is not from
self the expose chain will be walked.

Change-Id: I79c6a17df8581620b59e31b13c410f3a4053ec70