[benchmarks] Add SL4F wrapper for garnet_input_latency_benchmarks
This is based on the invocation in peridot/tests/benchmarks/benchmarks.cc,
which the SL4F-based version will eventually replace.

Change the test's shell script to be more aggressive about cleaning up
processes for the two reasons described in the comment.

This test kills Scenic, which would break screen_is_not_black_test,
which assumes that Scenic and the terminal (run on startup in the
"terminal" product) are running.  To address that, change
screen_is_not_black_test to reboot Fuchsia, so that it always tests
the initial startup condition.

Testing:
 * on CQ:
    * terminal-x64-release trybot for this new SL4F case
    * peridot-perf trybot for the old (non-SL4F) case
 * local testing:
   "fx set terminal.x64",
   "fx run-e2e-tests garnet_input_latency_benchmarks_test",
   "fx run-e2e-tests screen_is_not_black_test"

Bug: 23091
Change-Id: I768ecfc489a21957b647b96db0b1ae40ac993740