[rust-crates][pkg-resolver] switch resolver to using pin-project
The rust crate pin-project provides a safe mechanism for projecting
pinned values of a struct. This is safer than using pin-utils, because
it is impossible to cause undefined behavior (unless you use `unsafe`):

https://github.com/taiki-e/pin-project/pull/18

Additionally this switches the package resolver to using pin-project.

This project was approved by OSRB fxb/43118.

TEST: this is not expected to change behaivor, and should be covered by
the package resolver tests.

Fixed: 43878

Change-Id: Ib1bdd08710318e0bdaa983dc0e676b93b1139cd4