[audio_core] RouteGraph receives unique_ptr
RouteGraph now receives Renderers and Capturers as unique_ptrs. It
converts the pointers into fbl::RefPtr after receiving them to maintain
compatability in this CL with the AudioLink system.

The AudioLink system is removed in a following CL

=

This is part of a series of changes which remove AudioLink from AudioCore.
AudioLink requires an implicit distributed ownership model where
destination objects own sources, and sources own destinations, requiring us
to manually free objects.

The new feature pressuring this change is the addition of a
LoudnessTransform to links, which must be added to links by the RouteGraph.
The transform determines what loudness stages take effect for a given
output.

The replacement is the LinkMatrix, a single owner for all AudioObject
connections. RouteGraph will own the LinkMatrix and all AudioObjects.

Initially this refactor will not be strict about ownership, preferring
shared_ptrs.

Once some restructuring in audio core is done to ensure drop order
consistency across threads, more strictness can be imposed at compile
time.

Bug: 39624
Bug: 39426
Bug: 43514

TEST: changed; fx run-test audio_core_unittests

Change-Id: I04484831a6f438f3f6dfd89160e777a04554e95b