[components] Fix Breakpoints Flake
Prior to this CL, the root realm resolved receiver was installed on first
access to the breakpoint system which is too late.

Between the time the BuiltinEnvironment is setup and the test accesses the
breakpoint system, the root realm might've already been resolved, causing the
Blackbox test to hang.

The solution is to set up the breakpoint immediately when setting up the
BuiltinEnvironment. and passing along the receiver to the ScopedBreakpointSystem.

Bug: 43738
Change-Id: I0283233a42a8ef55ade738bab6ec9597b2802701