[audio_core] fbl::RefPtr -> std::shared_ptr
This changes uses of fbl::RefPtr for AudioObject and its descendents to
std::shared_ptr, so that we can use std::weak_ptr and custom deleters.

=

This is part of a series of changes which remove AudioLink from AudioCore.
AudioLink requires an implicit distributed ownership model where
destination objects own sources, and sources own destinations, requiring us
to manually free objects.

The new feature pressuring this change is the addition of a
LoudnessTransform to links, which must be added to links by the RouteGraph.
The transform determines what loudness stages take effect for a given
output.

The replacement is the LinkMatrix, a single owner for all AudioObject
connections. RouteGraph will own the LinkMatrix and all AudioObjects.

Initially this refactor will not be strict about ownership, preferring
shared_ptrs.

Once some restructuring in audio core is done to ensure drop order
consistency across threads, more strictness can be imposed at compile
time.

Fixed: 39624
Bug: 39426
Bug: 43514

TEST: no behaviour changed;
      fx run-test audio_core_unittests
      fx run-test audio_fidl_tests
      fx run-test audio_pipeline_tests
      fx run-test audio_fidelity_tests
      fx run-test audio_device_tests

Change-Id: I724956aa78cfe58d52a17b0e84906ed724a78a84