[serial] Introduce serial.Server
This is an abstraction that allows use to delegate reads and writes
to/from serial in a process separate from the one that opened the port.
Today botanist is responsible for managing serial i/o; however, we have
use-cases in which we'd also like for its subprocesses to read from and
write to serial, respectively (for running tests). This is a solution to
that problem.

Tested: unit tests and local integration tests.

Bug: 41377
Bug: 41930

Change-Id: I3be255325b62ae60828a49b10c0d3964455a569f