[ssh][syslog] Update SSHRunner and integrations
With this change:
* SSHRunner holds onto a SSH client instead of a single session - so that
it can run commands repeatedly. (Before it could only run one thing (as
sessions can only be Run() once), which was a faulty implementation.)
* SSHRunner can surface connection issues - and also gives the user the
ability to refresh the underlying connection in the event that the
client becomes unresponsive.
* Syslogger takes advantage of the above features and is able to recover
streaming in the event that the connection drops out, starting from that
point in time on reconnection (by default it starts streaming from t =
0).
* SSHTester was gutted and is now a thin wrapper around SSHRunner

SSH connection logic is difficult to unit test - and the cost of doing
so is high (i.e., an in-memory SSH server) . For example, see
https://github.com/golang/crypto/blob/master/ssh/session_test.go. Unit
tests however were added to the syslog integration. Moreover, the SSH
connection logic is thoroughly integration tested on every builder.

Also, this was manually tested with an internal test that reboots the
device:
https://chrome-swarming.appspot.com/task?id=498144625e757b10
syslog.txt was not cut short.

Change-Id: I9a7ebe4274ea211b81f3ec54506ac5cf55ce392f