[fidl][walker] Don't crash on primitive types.
The FIDL walker, used by the C family of encoders, decoders, and
validators, was asserting on primitive types. This was fine in previous
versions of coding tables that never emitted primitive types, but we now
emit a lot more type information in the coding tables for the
transformer, and it is dangerous to simply crash if a primitive type is
ever visited.

Test: runtests -t fidl-test  # added validate_primitives_struct test
Test: manually updated coded_types_generator.cc to output primitive types instead of nullptr for union fields, and verified that system boots and fx run-test fidl-test passes.

Change-Id: I1207d7b5baff44e02254142f8454b23f9f703f0c