[testrunner] overhaul
- testrunner was generally tidied and updated for style.
- "outputs" package was squashed into testrunner's main package; it gave
a confusing (and rather thin) layer of indirection.
- FuchsiaTester and SSHTester were squashed into FuchsiaSSHTester.
- a small amount of room was left open for the upcoming introduction of
FuchsiaSerialTester as well as the deprecation of out.tar.

Bugs fixed:
- we were invoking runtests without specifying a directory, which meant
that it was globbing a large number of directories on each invocation,
which is quite slow (and possibly buggy) on fuchsia
- data sinks are now being properly specified and added to the archive
- we no longer attempt to execute mac tests on linux and vice-versa.
- we were not previously running tests from the runtests package (as we
forgot to hook it up to the GN graph).
-tarutil.TarReader was renamed to the clearer tarutil.TarFromReader, and
it was given a far more efficient implementation.

Bug: 41930
Bug: 43500

Change-Id: I36992cd658d4a8ac9334b286ed2d2ccd464a2765