[lib][affine][test] Don't ASSERT_DEATH when debug assert disabled
Some of affine's tests assert that constructing a Ratio with a zero
denominator or inverting one with a zero numerator will terminate the
process.  These asserts are only enabled when
ZX_DEBUG_ASSERT_IMPLEMENTED is true.  Update the tests to ASSERT_DEATH
only when ZX_DEBUG_ASSERT_IMPLEMENTED is true.  This wasn't noticed
before because we rarely build with debug asserts disabled.

Bug: 42977
Test: /boot/test/sys/affine-test passes when assert_level = 0
Change-Id: I1a3ad03663de911e17013aa1b0175d5e2d98ff70