[components] Breakpoints: Introduce Injector trait
This CL introduces a new Injector trait for injecting capabilities
in Blackbox tests similar to the Interposer trait. This introduces
symmetry between the two APIs and will simplify installing Injectors
and Interposers ahead of time in the breakpoint system.

In a subsequent CL, the breakpoint system will get an install API
that accepts Interposers and Injectors and automatically installs them
when the matching capabilities are requested by a component.

Change-Id: I00d1eb8f29b565878b33b529985b6b825ae5f92e