[kernel][semaphore] Rework kernel Semaphores.
Change the implementation, and some of the semantics, of kernel
semaphores in an attempt to make their invariants easier to assert.

Previously, the semantics of a zircon kernel semaphore were that they
maintained a count which was supposed to be non-negative when there
were no waiters, and negative when there were.  Posting threads would
unconditionally increment the count, and then attempt to wake one
thread if the count had gone from negative to non-negative.  Waiting
threads would unconditionally decrement the count, then block if the
count had gone from non-negative to negative.

This approach needs to handle an edge case when a thread unblocks from
a wait operation due to a signal instead of a post operation.  In this
situation, the thread needs to restore the count it had previously
decremented, since a failed wait operation should not alter the
semaphore count.  When this happens, there is another chance for the
count to transition from negative to non-negative, mandating an
attempt to wakeup another thread.  See bug 42191 for an example of how
this can happen, and what can happen if the edge case gets missed.

As part of investigating this bug, I was trying to come up with a
simple invariant which can be asserted with minimal computation, and
it turns out to be more difficult then I had thought.  Ideally, we
could define C(S) as the count of semaphore S, and W(S) as the number
of waiters in S's wait queue, and then just say

  C(S) + W(S) >= 0

Unfortunately, when a thread times out of a wait queue, it is removed
from the wait queue, but may not immediately be scheduled.  It is no
longer part of the W(S) term given above because it is no longer in
the wait queue, it has been moved to the run-able queue but has not
been scheduled yet.  One can fix the invariant by adding a E(S) term,
where E(S) is the count of threads which have error'ed out of the wait
queue, but have not restored their previous count decrement operation.
The invariant would now be

  C(S) + W(S) + E(S) >= 0

But, this is not something we can easily compute while inside of the
thread lock to assert.  When a thread has left the wait queue, but not
made it back to the Semaphore::Wait operation, we have no way to track
it, so no good way to compute E(S), at least not without introducing
extra bookkeeping and overhead.

To complicate matters further, semaphores let you specify an initial
count, including an initial count which was negative.  Creating a
semaphore with a negative initial count would immediately violate the
invariant as stated.  Again, without extra bookkeeping, there is no
good way to construct an invariant which can be asserted.

So - to make this all a bit easier to deal with, we change the
semaphore implementation and semantics.  The API changes are as
follows.

++ Semaphores no longer have a count which can go negative.  Instead,
   the internal count is strictly unsigned.
++ Semaphores can still specify an initial count, but it must be
   unsigned.
++ Post no longer returns a signed post-Post-operation count where a
   negative count indicated the number of threads currently involved
   in a Wait operation.  Instead, Post now returns void.  count() and
   num_waiters() accessors were added to fulfill the testing purposes
   that Post's return value used to serve.

No one seemed to be using the ability to create a semaphore with a
negative count (in fact, no one was even using the ability to create a
semaphore with a non-zero initial count).  Also, no one (except tests)
ever paid any attention to the result of a Post operation.

Now that the count cannot be negative, Post becomes just

  if (Q.HasWaiters()) { Q.WakeOne(); }
  else { count++; }

and Wait just becomes

  if (count > 0) { count--; return OK; }
  else { return Q.wait(deadline); }

Even though a thread may become signaled and leave the wait queue
without having made it back to the Wait method, it cannot leave the
wait queue at all while the thread lock is held.  This allows us to
assert a simple invariant at any time that we are inside of the thread
lock.  This is:

  (C(S) == 0) || (W(S) == 0)

It is OK for a semaphore to have neither waiters, nor a non-zero
count, but it cannot have both waiters *and* a non-zero count.

Finally, the existing tests exercised a number of different cases, but
for some reason, they never attempted to actually wake a blocked
thread using a Post operation.  While updating the tests, fix this
oversight.

Summary:
++ Change the semaphore API to disallow negative counts.
++ Remove the return value from Post.
++ Add count and num_waiters accessors to allow for tests.
++ Update the tests.
++ Add the wake-via-post test.
++ Change the internal implementation, and assert the new invariant
   during post and wait operations.

Fixed: 42191
Change-Id: I8482853cd324a0ab7a506e600c5228e2cbafa94a