[fidlcat] create values from syscalls
The goal is to have only one unify representation for syscalls and FIDL
messages.

This way it's easier to serialize, compare, print, ...

For syscalls which can be converted to values, when the syscall is
decoded, we create the values and then print these values. For the other
syscalls, we directly print the syscall.

This patch only create the minimal working implementation.

It only works for inline input arguments which are all handles.

The flags fidl_codec_values_ready_ tells if the syscall input arguments
can be converted to values. When everything will be implemented, this
flag will disapear (it would be always true) and the code which prints
directly a syscall will be removed.

The current tests ensure that this new two step printing behaves exactly
like the direct printing.

Change-Id: Icb679408932b43d95612ea31c31abbc9ba08c309