[kernel][arm64] Remove the local cache invalidate in the top of start.S
This was never really safe, but the cache invalidate loop has been
broken for a while so it hadn't actually been doing anything.

Trouble is secondary cpus boot in through the same path and can trash
the L2/L3 caches of the existing cpus that have already been running.
This shows up as a few flakes as the system starts up on real hardware
with a change where the existing cache flush routine is fixed.

Assume the caches of all the cpus are in a consistent state when
entering the kernel, which should be safe. CPU0 will enter from the
previous bootloader which should leave the cpu in a good place and the
secondaries are started from PSCI which makes hard guarantees that the
cache state of secondaries as they're brought up are clean and coherent.

Tested: boot and zedboot on sherlock as well as plain qemu and validate
that everything seems stable. Run a bunch of CQs to make sure it doesn't
introduce a new flake.

Change-Id: Ifd95c4a00568e9da327b817c6cc76a1f02870719