[kernel][arm64][debug] Fix TestDebugRegistersValidation core test
Fix the deterministic failures of TestDebugRegistersValidation.  Prior
to this change, the test deterministically failed in two ways:

1. It expected that setting more breakpoints than the platform
allowed, would result in an error.  However, the comment for
hw_bps_count is debug.h indiciates that the count is ignored when
written.

2. It expected that when a breakpoint was disabled, the DBGBCR would
still contain BAS and PMC values rather than 0.

This change fixes #1 by removing the check from the test.

This change fixes #2 by clearing the entire DBGBCR when the user has
not set the E bit.

Note, this test is known to be flaky and this change does nothing to
address that.  That'll get fixed in a future change.

Bug: 33942
Test: temporarily enabled test and ran on Cavium ThunderX2
Change-Id: Iff350c919f2ae3b7c1d33511605a836df0420da5