[wlan][mlme] handle_mlme_scan_req -> on_sme_scan
Keeping the handle_mlme_msg unchanged because MLME message is a valid
term, meaning a message defined in the MLME protocol.

on_sme_scan is an action that is carried out per SME's instruction.
Based on the pattern in the MLME protocol, everything from SME is a
request so the _req part is left out for simplicity.

Bug: 29663
Test: Naming change does not change functionality.
Change-Id: I621745af8716b5b017607453177ca75bf28b3e9c