[activity] Add fuchsia.ui.activity.* to SDK
fuchsia.ui.activity allows clients to notify the Activity Service of
user activity, and subscribe for notifications of the activity state
(IDLE/ACTIVE) of the device changing based on user activity.

fuchsia.ui.activity.control is a transitional API that allows clients to
push device idle state to the Activity Service, overriding the Activity
Service's internal determination of the device's activity state.

Bug: 34734
Change-Id: Iae5aa09e7d5de4c29884935eea0247a0b0f14985