[debugger] Remove "edit" command.
This was the way to edit breakpoints. We will be moving to using the
setting system to edit breakpoint attributes and the old structure is
making it difficult to change the breakpoint code. Since the uses of
edit were obscure, we should be fine without this capability for a
period.

Change-Id: I9a90933cc0031e7b35297200307c962101838401