[SetUI] Revise Agent Interaction.
Setting Service Agents were originally designed to be flexible,
passing a sender to the Agent Authority to facilitate interactions.
However, the authority exclusively controls the agents and therefore
has a natural ownership of the their lifespans. On the other hand,
the registration process was excessively rigid. An agent could only
register to service a single lifespan, while implementations can call
for grouping together the logic of multiple lifespans.

This change corrects these issues by introducing Agent, a trait
that can be implemented and handed into the Agent Authority. To
address multiple lifespans, it is now left to the Agent to reject
lifespans that are not desired.

Bug: 43884
Test: The majority of these changes are a refactor of existing logic
and can be tested with the present agent tests. These tests have been
revised to use the new interfaces/interactions. Tests no longer  pass
agents around as the Agent Authority can and does lock the agents
during invocation.

fx run-test setui_service_tests -- --test

Change-Id: I40b6f8dff32fa676829d5c3e8dd336e6b521e484