[media][audio] Fix calling DiscardAllPackets before configured
When DiscardAllPackets was called on an AudioRenderer before the
configuration had been validated, it would result in a timeline function
being set for the renderer that was stopped. Later calls to
Play/SendPacket would not move the timeline forward because the
transform was already marked as valid.

This change does not update the timeline function for the renderer
unless the configuration has been validated first.

Test: fx run-test audio_fidl_tests
Bug: 43508
Change-Id: Ic3d05a1baad72a6f43a5487be1c0153d2e8fa1e6