[brcmfmac] Add intrusive_list<>
intrusive_list<> is an intrusive doubly-linked list implementation in
the spirit of Linux's "struct list_head".  It has the the useful
property in particular of auto-removal from lists when an element is
independently destroyed, and is notably more lightweight than the FBL
intrusive list implementation.  It also attempts to hew closely to STL
conventions, where applicable.

Bug: 40719
Test: Added a test to, and executed "fx run-test brcmfmac_test"

Change-Id: I22c9ca895dca1a80aed1c26f979b1276679cfd64