[debugger] Parse expressions to InputLocations.
Adds a new variant to parse InputLocations that can handle full
expressions. This adds complexity because expression evaluation may be
asynchronous. This is done as a new function but to allow an easier
transition and because some callers don't need expression evaluation
(e.g. "until").

Some additional work will need to be done in the ResolveUnique*() calls
to support expressions in a followup.

Change-Id: I6f94ed1577c1beca1dffc0e58732848b78379126