[fidlcat] values for integers
With this patch, we are able to decode input arguments for
zx_channel_create and zx_channel_read using the new mechanism.

There is no visible change (which means that the tests are still working).

Handle value generation for Bool and integer types.

Add a hexadecimal_display flag to the integral types.

Fix zx_ioports_release to be consistent with zx_ioports_request (now
both io_addr field types are Uint16Hexa).

Change-Id: I5bdf9dd4f8359d6a97f02b2bc88bc3ac583a528b