[sfw][input] Add InputHandler and ImeHandler
An InputHandlers filter and dispatch InputEvents to relevant services.

This change adds the ImeHandler, which filters on KeyboardEvents and
dispatches them to IME. Since the Scenic session is currently still
handling keyboard events, the ImeHandler also sends KeyboardEvents to
Scenic, but this will be removed once we're transitioned to input2
APIs.

TEST = fx run-test input_tests

Change-Id: I43b557affd4396268fe56e874477f0469603c486