[fuchsia-pkg-testing] Speed up blob reads
When verifying large packages, it is faster to first get a VMO of the
file and use that to obtain a byte buffer of the file.  This change
modifies verify_package's read_file API to first try obtaining a VMO
version of the file and falling back to `read` if that fails.

On a debug asan build running on a NUC, this reduces the total test
runtime from 181s to 131s (a 28% reduction).

Test: pkg-resolver-integration-tests
Bug: 43380
Change-Id: Iac2bf50562010d8b9cf526078bf7f0a61f9e7da0