[fidlc] Avoid generating duplicate coding tables
Duplicate coded types would be generated, causing invalid C to be
generated for the coding tables. There was a bug where if there was an
array type whose member type needed a coding table (all non-array,
non-primitive types) and it was used in both an envelope and
non-envelope context then two coding tables would be generated.

This changes the way that arrays and primitives are tracked in the coded
types generator. Previously they were tracked by a pair of
(context, type), but that was insufficient. Now they're tracked by a
pair of (coding_needed, type).

Change-Id: Ife37557c2921fe868692101290935a8535c5ae9c