[debugger] Expose breakpoint attrs in settings.
Allows "get" and "set" to be used for all breakpoint attributes except
for the type (which will be done in a followup).

Moves around some string processing code for formatting stop modes and
booleans so it can be better shared.

Uses setting names as the headings for the breakpoint table so these are
consistent.

Uses setting names for breakpoint switches for consistency. Remove the
--enable switch from breakpoint and replace it with --disabled which is
a little simpler to type. Add an option to create one-shot breakpoints.

Change-Id: Ifb01af7de66df9c80c4fd8e9703913fe25818d8c