[debugger] Remove callback from Breakpoint.SetSettings.
The callback on BreakpointSetSettings complicated the code with
asynchronous failure handling. This isn't particularly helpful because
there's not much that the caller can do in response to such a failure,
because it's normally fatal.

Additionally, errors can occur at any time because processes and modules
can get added at runtime. So having the SetSettings callback is not
sufficient either.

The callback is replaced with a general breakpoint observer function
that will get called when needed to communicate the error to the user.

Change-Id: I2616c3b304012e9c1224209322daaa7dbce21853