[fonts] Add empty font collection for products that need it
Some headless products might depend on the font server due to technical
constraints, but don't actually need any fonts.

This is the premier font collection for those products.

Adding "//src/fonts:empty-font-collection" to a product's dependencies
will generate a valid but empty font manifest and provide it to the
font server.

Bug: 44158
Change-Id: Ib6b1549ebd1273f465937c24d70b2f82f195369e