[components] Correct path to component
In this test we're attempting to supply a root component manifest
that is invalid. Previously this specified an invalid URL. Now
the URL is valid and does what it intends to, which is point to
the component manager binary itself. Note: We are purposefully
pointing to a binary, not a manifest.

Change-Id: I33e04e2faf6992b1ed71c6904bf367e2aa30eff4