[benchmarks][scenic] Fix bad testing caused by missing package dependency
This fixes a regression introduced by this infra config change in
integration.git (only visible in the internal integration.git):
  [benchmarks] Build "terminal" instead of "core" on peridot-perf and perfcompare
  (I2dd963926a3c4c445a5b51df74dd8b983ac91382)

That change was not directly testable, but it was tested via this CL:
https://fuchsia-review.googlesource.com/c/fuchsia/+/350503

That change switched to a much-reduced set of package dependencies,
but it removed a package, set_renderer_params, used by the Scenic
benchmarks.  Before, the logs would contain this:

  == fuchsia.scenic.hello_scenic: Configuring scenic renderer params...
  [WARNING:src/ui/tools/set_renderer_params/main.cc(22)] This tool is intended for testing and debugging purposes only and may cause problems if invoked incorrectly.

After, they contained this:

  == fuchsia.scenic.hello_scenic: Configuring scenic renderer params...
  fuchsia-pkg://fuchsia.com/set_renderer_params#meta/set_renderer_params.cmx: not found

That means the variants of the Scenic tests were failing to set
different Scenic parameters, and so were failing to test the variants
they were intended to test.

This explains why the perfcompare trybot's results on
https://fuchsia-review.googlesource.com/c/fuchsia/+/350503 showed
performance differences on various Scenic tests!

I only noticed this cause when adding an SL4F wrapper for these tests
and seeing the error message.

This mistake is caught by enabling shell error handling, which
requires two changes:

 * Ignore errors from the killall commands using "|| true".  This
   matches garnet/tests/benchmarks/input_latency/*.sh, run by
   garnet_input_latency_benchmarks_test.

 * Add the missing package dependency, set_renderer_params.

Also, avoid using $1 when it might be undefined.  This happens not to
be covered by the test cases, because those exit the while loop via
the "break" case first.  However, it's better to change this, and it
also matches garnet/tests/benchmarks/input_latency/*.sh.

Bug: 23091
Test: on CQ: terminal-x64-release (newer, SL4F), peridot-perf (older, non-SL4F)
 * local testing: "fx set terminal.x64", "fx run-e2e-tests scenic_benchmarks_test"
Change-Id: Ie0d15a8e522a6b3ab717366bac09249f0e5a0621