[amlogic-decoder][h264] Fix DPB size calculation
We need to be more careful when trying to fix up the DPB size in the
face of invalid inputs, as otherwise we generate invalid values and hit
some asserts. Also some inputs in tests were hanging flakily because
there weren't enough buffers, so increase the limit in those cases and
add a new test to ensure we always test the minimum frame count.

Bug: 41492

Change-Id: I629654e7def288ad625892689b5c5a61b83dcf69