[botanist] Refactor
Botanist has grown unwielidy over the past ~year and so this change
returns it to a more manageable and scalable state. Much of the weight
loss is due to the removal of log support in the multi-device case, which -
interestingly - has not been used at all or relied upon since its introduction!
(Indeed, we are creating files (with %d suffixes that we have up front
not registered as task outputs, and so they never get shipped back to
the recipe).)

Also removed was botanist's care for not having streamed empty log
content. This seems like something that should more reasonably be
checked in the context of a recipe rather than something botanist should
intimately care about.

Tested: botanist is extensively integration tested being the tool that
CQ itself consumes from source to run tests and manage device
interaction. Also, this change can be viewed as bootstrapping
testability for a more streamlined botanist codebase.

Change-Id: I5809ec79e745339a4b970f8a109e2bb47fe2eef9