[escher] Fix Scenic GLSLang build error
This CL makes it easier for clients of Escher to
include it as a dependency without having to know
which version of Escher to use (e.g. with or without
glsl support). Clients can simply include "escher" and
the correct version will be included automatically (based
on fx set build flags).

Clients that always need a specific version of Escher can
instead depend on either "escher_with_glslang" or
"escher_no_glslang" depending on their needs.

Change-Id: I11f34c146d439f84dc4155a8aa7ba4e47e70fd76