[fidl_codec] remove more json field uses
Now we don't know the size of members for which we can't decode the
type. That means that we no longer have an hexadecimal dump of value of
unknown type.

To reflect that, we removed the inline_size to several type decoding
functions. We also renamed RawType into InvalidType.

Remove now unused methods.

Reorder type decoding status function in wire_types.h to have the same
order found in wire_types.cc

Bug: 43957

Change-Id: I6af2f45a023dc3c3e9518d44e180e766ceff0a07