[fidl][cpp] Simplifying code to destruct in unions
When two types named the same need to be destructed, the expression
`var.~Type()` is ambiguous and the compiler errors out. To avoid this
problem altogether, we are switching the generation to
`var.~decltype(var)`. This in turns reduces the need to track destructor
types in the fidlgen cpp IR code, which can be simplified to only track
whether explicit destruction is needed (or not).

Test: goldens, and other FIDL tests
Change-Id: I2289385af60b4aa87f83c73c45a4bbb1fd8f389a