[fidlc] Expose anonymous request/response structs in IR
These new structs (which represent request/response payloads)
will eventually be used as the source of truth for request/
response type/fieldshape info in fidlgen, which will allow us
to remove special handling of requests and responses.

In order to emit the request/response payloads as regular structs,
a temporary wire format kV1Header is added which is the same as
kV1NoEe except it does not treat request/response payloads specially
compared to regular structs (by ignoring the is_request_or_response field).

In addition, since both structs and requests/responses are generated
using the same Generate(flat::Struct) function in the json_generator,
this function now takes an extra bool to indicate whether its input
struct is a request or a response. This bool is set to true when called
from GenerateRequest in order to generate type info using the kV1NoEe
wire format. Otherwise, the kV1Header wire format is used.

Change-Id: I605fd790ef312e20842fc3574871d95276f1e68d