[build] Fix enable_netboot build argument
This is a reland of 5923ae243440f539fba228f80eda7a9c79d82249 with fixes.

At least two past changes made enable_netboot=false have no effect.
This means every build running the default target is doing a very
large and costly image construction/compression step it doesn't need.

In this change, the `archive = true` field in metadata.images becomes
mandatory for a file to be included in the build archive.  The netboot
file is only included in the archive if `enable_netboot` is set.

Fixed: 44352
Change-Id: I62e923bcde8cbc2150e02c820cc65f5e1cb28d2b