[benchmarks] Fix uses of "*" in some shell scripts that could go wrong
If 'killall scenic*' is run from a current directory that happens to
contain a file called scenic.txt (for example), it will expand to
'killall scenic.txt', which will fail to have the intended effect of
killing the scenic process.

Fix this by quoting the arguments so that the shell won't interpret
them as glob expressions.

Bug: none
Test: on CQ: terminal-x64-release and peridot-perf
 * local testing:
   "fx set terminal.x64",
   "fx run-e2e-tests garnet_input_latency_benchmarks_test",
   "fx run-e2e-tests scenic_benchmarks_test"
Change-Id: I227f643035db2e506df4a8e1e6acbf2194f78933