[fidl][go] Use struct typeshapes for request/responses
This updates fidlgen to use the payload structs instead of
request/response type info from thee IR.

In fidlgen, this means anonymous structs are compiled but stored
separately from the rest of the structs. They get their correct
name set in compileMethod, and then afterwards are appended to
Root.Structs.

Also, previously .Request/Response would be nil when a method does
not have a request/response, and would have empty .Members when
there was no payload. Now, .Request/Response is nil when there is
no payload, and .HasRequest/Response is used to indicate whether
there is a request/respones or not (if .Request/Response are not nil,
it must have at least one member). This makes it clear that
empty payloads are a different case from an empty struct payload.

Bug: 7704
Test: fx run-test go_fidl_tests
Change-Id: I395ce017b993dbac0c7306d2557ca05a5fa987bd